move indexedCollection
