+++ core/incremental/indexedCollection.ts
